requirements.txt
flask
ultralytics
pillow
werkzeug
